pyproject.toml
[project]
name = "game-nim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "mysql-connector-python>=9.3.0",
]
